requirements.txt
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorful==0.5.5
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
environs==9.5.0
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
google-api-core==2.14.0
google-auth==2.25.1
googleapis-common-protos==1.61.0
gpustat==1.1.1
grpcio==1.58.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
humanfriendly==10.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
minio==7.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
nvidia-ml-py==12.535.133
onnxruntime==1.16.3
opencensus==0.11.3
opencensus-context==0.1.3
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.19.0
protobuf==4.25.1
psutil==5.9.6
py-spy==0.3.14
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.8.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.6
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.21.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
yarl==1.9.3
zipp==3.17.0
